@@ -4,44 +4,46 @@
   "license": "MIT",
   "angular-cli": {},
   "scripts": {
-    "start": "npm run build-themes && ng serve",
-    "start-jit": "npm run build-themes && ng serve --aot=false",
+    "start": "yarn build-themes && ng serve",
+    "start-jit": "yarn build-themes && ng serve --aot=false",
     "lint": "tslint -p src/tsconfig.json",
-    "test": "npm run build-themes && ng test",
+    "test": "yarn build-themes && ng test",
     "pree2e": "webdriver-manager update",
     "e2e": "protractor",
     "build-themes": "bash ./tools/build-themes.sh",
-    "prod-build": "npm run build-themes && ng build --aot --prod",
-    "postinstall": "webdriver-manager update --gecko false && bash ./tools/build-themes.sh",
+    "prod-build": "yarn build-themes && ng build --aot --prod",
+    "postinstall": "webdriver-manager update --gecko false && yarn build-themes",
     "publish-prod": "bash ./tools/deploy.sh stable prod",
     "publish-dev": "bash ./tools/deploy.sh",
     "publish-beta": "bash ./tools/deploy.sh stable beta"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.4",
-    "@angular/cdk": "^7.2.1",
-    "@angular/common": "^7.1.4",
-    "@angular/compiler": "^7.1.4",
-    "@angular/core": "^7.1.4",
-    "@angular/forms": "^7.1.4",
-    "@angular/http": "^7.1.4",
-    "@angular/material": "^7.2.1",
-    "@angular/material-examples": "https://github.com/angular/material2-docs-content.git",
-    "@angular/material-moment-adapter": "^7.2.1",
-    "@angular/platform-browser": "^7.1.4",
-    "@angular/platform-browser-dynamic": "^7.1.4",
-    "@angular/router": "^7.1.4",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.1.4",
+    "@angular/cdk-experimental": "^8.1.1",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/material": "^8.1.4",
+    "@angular/material-examples": "angular/material2-docs-content#8.1.x",
+    "@angular/material-experimental": "^8.1.1",
+    "@angular/material-moment-adapter": "^8.1.4",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "core-js": "^2.6.1",
     "hammerjs": "^2.0.8",
+    "material-components-web": "^1.1.1",
     "moment": "^2.23.0",
-    "rxjs": "^6.3.3",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.1",
+    "zone.js": "^0.9.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.11.4",
-    "@angular/cli": "^7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
+    "@angular-devkit/build-angular": "^0.800.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
     "@types/jasmine": "^3.3.5",
     "@types/node": "^10.12.18",
     "firebase-tools": "^6.2.2",
@@ -56,6 +58,6 @@
     "protractor": "~5.3.0",
     "ts-node": "^6.0.3",
     "tslint": "^5.10.0",
-    "typescript": "~3.1.3"
+    "typescript": "3.4"
   }
 }
