@@ -9,7 +9,7 @@ import {
   ViewChild,
   ViewEncapsulation,
 } from '@angular/core';
-import {MatTabsModule} from '@angular/material';
+import {MatTabsModule} from '@angular/material/tabs';
 import {ActivatedRoute, Params, Router, RouterModule} from '@angular/router';
 import {combineLatest, Observable, Subject} from 'rxjs';
 import {map, takeUntil} from 'rxjs/operators';
@@ -35,17 +35,21 @@ export class ComponentViewer implements OnDestroy {
               public _componentPageTitle: ComponentPageTitle,
               public docItems: DocumentationItems,
               ) {
+    let params = [_route.params];
+    if (_route.parent) {
+      params.push(_route.parent.params);
+    }
     // Listen to changes on the current route for the doc id (e.g. button/checkbox) and the
     // parent route for the section (material/cdk).
-    combineLatest(_route.params, _route.parent.params).pipe(
+    combineLatest(params).pipe(
         map((p: [Params, Params]) => ({id: p[0]['id'], section: p[1]['section']})),
         map(p => ({doc: docItems.getItemById(p.id, p.section), section: p.section}),
         takeUntil(this._destroyed))
         ).subscribe(d => {
-          this.componentDocItem = d.doc;
-          if (this.componentDocItem) {
+          if (d.doc !== undefined) {
+            this.componentDocItem = d.doc;
             this._componentPageTitle.title = `${this.componentDocItem.name}`;
-            this.componentDocItem.examples.length ?
+            this.componentDocItem.examples && this.componentDocItem.examples.length ?
                 this.sections.add('examples') :
                 this.sections.delete('examples');
           } else {
@@ -65,8 +69,8 @@ export class ComponentViewer implements OnDestroy {
   encapsulation: ViewEncapsulation.None,
 })
 export class ComponentOverview implements OnInit {
-  @ViewChild('initialFocusTarget') focusTarget: ElementRef;
-  @ViewChild('toc') tableOfContents: TableOfContents;
+  @ViewChild('initialFocusTarget', {static: false}) focusTarget: ElementRef;
+  @ViewChild('toc', {static: false}) tableOfContents: TableOfContents;
   showToc: Observable<boolean>;
 
   constructor(public componentViewer: ComponentViewer, breakpointObserver: BreakpointObserver) {
