@@ -7,10 +7,10 @@ import {DocsAppTestingModule} from '../../testing/testing-module';
 const guideItemsId = 'getting-started';
 
 const mockActivatedRoute = {
-  fragment: Observable.create(observer => {
+  fragment: new Observable(observer => {
     observer.complete();
   }),
-  params: Observable.create(observer => {
+  params: new Observable(observer => {
     observer.next({id: guideItemsId});
     observer.complete();
   })
