@@ -1,13 +1,11 @@
 import {DocViewer} from './doc-viewer';
 import {ExampleViewer} from '../example-viewer/example-viewer';
 import {StackblitzButtonModule} from '../stackblitz/stackblitz-button';
-import {
-  MatButtonModule,
-  MatIconModule,
-  MatTabsModule,
-  MatTooltipModule,
-  MatSnackBarModule,
-} from '@angular/material';
+import {MatButtonModule} from '@angular/material/button';
+import {MatIconModule} from '@angular/material/icon';
+import {MatSnackBarModule} from '@angular/material/snack-bar';
+import {MatTabsModule} from '@angular/material/tabs';
+import {MatTooltipModule} from '@angular/material/tooltip';
 import {PortalModule} from '@angular/cdk/portal';
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
