@@ -428,14 +428,22 @@ const DOCS: {[key: string]: DocCategory[]} = {
           name: 'Drag and Drop',
           summary: 'Directives enabling drag-and-drop interactions',
           examples: [
-            'cdk-drag-drop-axis-lock',
-            'cdk-drag-drop-sorting',
             'cdk-drag-drop-overview',
-            'cdk-drag-drop-horizontal-sorting',
-            'cdk-drag-drop-custom-preview',
-            'cdk-drag-drop-root-element',
-            'cdk-drag-drop-handle',
+            'cdk-drag-drop-axis-lock',
+            'cdk-drag-drop-boundary',
             'cdk-drag-drop-connected-sorting',
+            'cdk-drag-drop-connected-sorting-group',
+            'cdk-drag-drop-custom-placeholder',
+            'cdk-drag-drop-custom-preview',
+            'cdk-drag-drop-delay',
+            'cdk-drag-drop-disabled',
+            'cdk-drag-drop-disabled-sorting',
+            'cdk-drag-drop-enter-predicate',
+            'cdk-drag-drop-free-drag-position',
+            'cdk-drag-drop-handle',
+            'cdk-drag-drop-horizontal-sorting',
+            'cdk-drag-drop-root-element',
+            'cdk-drag-drop-sorting'
           ],
         },
         {
@@ -492,7 +500,14 @@ const DOCS: {[key: string]: DocCategory[]} = {
         {id: 'stepper',
           name: 'Stepper',
           summary: 'Presents content as steps through which to progress.',
-          examples: []
+          examples: [
+            'stepper-vertical',
+            'stepper-editable',
+            'stepper-optional',
+            'stepper-errors',
+            'stepper-label-position-bottom',
+            'stepper-states'
+          ]
         },
         {
           id: 'table',
@@ -528,8 +543,9 @@ for (let category of DOCS[CDK]) {
 }
 
 const ALL_COMPONENTS = DOCS[COMPONENTS].reduce(
-  (result, category) => result.concat(category.items), []);
-const ALL_CDK = DOCS[CDK].reduce((result, cdk) => result.concat(cdk.items), []);
+  (result: DocItem[], category: DocCategory) => result.concat(category.items), []);
+const ALL_CDK = DOCS[CDK].reduce(
+  (result: DocItem[], cdk: DocCategory) => result.concat(cdk.items), []);
 const ALL_DOCS = ALL_COMPONENTS.concat(ALL_CDK);
 const ALL_CATEGORIES = DOCS[COMPONENTS].concat(DOCS[CDK]);
 
@@ -549,12 +565,12 @@ export class DocumentationItems {
     return [];
   }
 
-  getItemById(id: string, section: string): DocItem {
+  getItemById(id: string, section: string): DocItem | undefined {
     const sectionLookup = section == 'cdk' ? 'cdk' : 'material';
     return ALL_DOCS.find(doc => doc.id === id && doc.packageName == sectionLookup);
   }
 
-  getCategoryById(id: string): DocCategory {
+  getCategoryById(id: string): DocCategory | undefined {
     return ALL_CATEGORIES.find(c => c.id == id);
   }
 }
