@@ -1,5 +1,5 @@
 import {Component, Input} from '@angular/core';
-import {MatSnackBar} from '@angular/material';
+import {MatSnackBar} from '@angular/material/snack-bar';
 import {ComponentPortal} from '@angular/cdk/portal';
 
 import {EXAMPLE_COMPONENTS, LiveExample} from '@angular/material-examples';
