@@ -1,7 +1,9 @@
 import {Component, Input, NgModule} from '@angular/core';
 import {StackblitzWriter} from './stackblitz-writer';
 import {ExampleData} from '@angular/material-examples';
-import {MatButtonModule, MatIconModule, MatTooltipModule} from '@angular/material';
+import {MatButtonModule} from '@angular/material/button';
+import {MatIconModule} from '@angular/material/icon';
+import {MatTooltipModule} from '@angular/material/tooltip';
 
 @Component({
   selector: 'stackblitz-button',
@@ -16,7 +18,7 @@ export class StackblitzButton {
    * The button becomes disabled if the user hovers over the button before the stackblitz form
    * is created. After the form is created, the button becomes enabled again.
    * The form creation usually happens extremely quickly, but we handle the case of the
-   * stackblitz not yet being ready for people will poor network connections or slow devices.
+   * stackblitz not yet being ready for people with poor network connections or slow devices.
    */
   isDisabled = false;
   stackblitzForm: HTMLFormElement;
