@@ -6,15 +6,13 @@ import {
   OnInit,
   OnDestroy,
 } from '@angular/core';
-import {StyleManager} from '../style-manager/style-manager';
+import {StyleManager} from '../style-manager';
 import {ThemeStorage, DocsSiteTheme} from './theme-storage/theme-storage';
-import {
-  MatButtonModule,
-  MatGridListModule,
-  MatIconModule,
-  MatMenuModule,
-  MatTooltipModule,
-} from '@angular/material';
+import {MatButtonModule} from '@angular/material/button';
+import {MatGridListModule} from '@angular/material/grid-list';
+import {MatIconModule} from '@angular/material/icon';
+import {MatMenuModule} from '@angular/material/menu';
+import {MatTooltipModule} from '@angular/material/tooltip';
 import {CommonModule} from '@angular/common';
 import {ActivatedRoute} from '@angular/router';
 import {Subscription} from 'rxjs';
@@ -65,7 +63,10 @@ export class ThemePicker implements OnInit, OnDestroy {
     public styleManager: StyleManager,
     private _themeStorage: ThemeStorage,
     private _activatedRoute: ActivatedRoute) {
-    this.installTheme(this._themeStorage.getStoredThemeName());
+    const themeName = this._themeStorage.getStoredThemeName();
+    if (themeName) {
+      this.installTheme(themeName);
+    }
   }
 
   ngOnInit() {
